fix(assertions): nested stacks inside non-root stages don't resolve templates
Templates are placed inside a path described by their closest assembly. Using this assembly lets nested stack templates resolve, regardless of their stage depth.

I took the opportunity to reduce the scope of synth to only the containing template, regardless of the stack being nested or not. This should decrease load on user machines when synthesizing templates for apps with multiple stages.

Closes #24004.