fix(pipelines): cannot configure actionName for all sources
Add the ability for GitHub, CodeStar Connections, ECR source types
to configure the Action Name that is added to the pipeline.

This is necessary to add the same GitHub repository to the same pipeline
twice with a different branch each time. By default, it would pick the
same default name for each, which would then conflict.