fix(cdk-assets):  asset concurrency leaves a corrupted archive (#24026)
This is a re-roll of https://github.com/aws/aws-cdk/pull/23677 which was reverted in https://github.com/aws/aws-cdk/pull/23994 because the `randomUUID()` function from the original solution was not available in Node versions below 14.17 (and we advertise compatibility with Node 14.*).

We didn't actually need a UUID, just any random string, so replace it with a function that generates a random string in a different way.

----------

Resolves #23290

A very simple fix for the issue where builds with `--concurrency` specified can lead to corrupt archives. Rather than use the outputFile as the basis for the temp file name we simply use a random string.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*