chore(pipelines): write a GraphViz file with the pipeline structure (#24030)
This is a re-roll of https://github.com/aws/aws-cdk/pull/23908 which had to be reverted in  https://github.com/aws/aws-cdk/pull/24006 because there are some cases where change set approval steps that are shared between two stacks would cause cyclic dependencies between those stacks that caused them to be mutually unsortable.

Solve this by adding a mode to the toposort routine that will proceed even if there are cyclic dependencies (which is used purely for rendering).

----

Add a `pipeline.dot` file to the cloud assembly containing the graph structure of the pipeline.

This change is a `chore`, not a `feat`, as I want this to be a debugging aid, but I don't want to service feature requests on it.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*