fix(pipelines): cannot configure actionName for all sources (#24027)
Add the ability for GitHub, CodeStar Connections, CodeCommit source types to configure the Action Name that is added to the pipeline.

This is necessary to add the same GitHub repository to the same pipeline twice with a different branch each time. By default, it would pick the same default name for each, which would then conflict.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*