fix(codepipeline): x-acct/x-region ECS deployment misses support stack dependency
In the case of a cross-account AND cross-region ECS deployment, a
dependency between the support stacks that is necessary is missing.

This cannot be fixed globally. Because the resources reference each
other bidirectionally (user -> bucket, bucket -> user), the only way to
fix this is to do it locally: in the CodePipeline module, where we can
have the knowledge that we use generated names and that everything will
work out if we deploy the role before the bucket.

All CodePipeline Actions must have this fix eventually, but since
people may have crazy stack setups in which addition of this dependency
may introduce a cyclic dependency (breaking the synth), we're rolling
this fix out with limited blast radius.

Follow-up in #24050, and suggest a good clean-up in #24051.