fix(s3): infer bucketWebsiteUrl and bucketDomainName suffixes from bucket region (#23919)
The current implementations of determining various bucket URLs was incorrectly relying on the stack's region.
In practice this did not matter a lot, since the suffix depends on the partition and a cross over from one partition to another is not likely or even impossible.

We also required users to determine the correct bucketWebsiteUrl format for a bucket.
However this information can reliably be inferred from a bucket's region.
Since bucket ARNs do not contain region information, it is best to provide the bucket region whenever bucketWebsiteUrl will be used.
As a fallback we assume a bucket is placed in the region of its stack.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license