fix(core): cross-stack reference error doesn't include violation (#23987)
When users are using cross-stack references that are not allowed, we just tell them the stacks that are involved in the reference, but not the reference itself. For example:

```
Stack "ExampleStack" cannot consume a cross reference from stack
"cross-region-stack-1111111111:us-east-2". Cross stack references are
only supported for stacks deployed to the same environment or between
nested stacks and their parent stack.
```

This makes it very hard to debug what's going on, and why the reference is there in the first place.

Render the reference as well so that it's easier to figure out why this is happening.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*