fix: Use the correct LB full name when creating metrics for imported LBs (#23972)
The `Arn.split()` method doesn't parse the `resourceName` correctly when it has multiple `/`, which is the case for the resources created by by the elbv2 API.

I've also refactored the `integ.nlb-lookup.ts` test because it was not well written and I couldn't really deploy it with Cloudformation.

The capability to create metrics from imported Load Balancers is new. It was introduced in #23853.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*