chore(cli-integ): try to fix some integ test failures
Working on fixing some cli-integ test failures that we see. Some of the
failures seem to relate to the bootstrap tests, and specifically the
tests that have S3 assets (i.e. Lambda functions). For some reason when
the Lambda is deployed the asset does not exist in the bucket.

In this PR i've made a couple of updates.

1. Ensure that the bootstrap tests do not also run `ensureBootstrapped`
   since the test itself is responsible for bootstrapping.
2. Ensure that the correct context is passed to each test

re https://github.com/cdklabs/cdk-ops/issues/2341