chore: add labels to pr based off comment content and labels already added (#24095)
This will help us find the PRs that need attention even when they don't pass the PR Linter, the build, and/or already have changes requested.

The GitHub action adds a label for cli tests needed, for clarification requested, and for exemptions requested.

The PR linter output has been updated with a message to let contributors know to use these. I added in both the review comment and the regular comment so that it's not easy to miss.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*