chore: add TimeZones class (#24127)
This adds the time zones in the 2022g release of the Time Zone Database.

It has been tested in the context of the tests written in https://github.com/aws/aws-cdk/pull/23123, but I didn't commit them here because I didn't want to overwrite that change.

I couldn't decide whether this should be a chore or a feat. I'm open to either.

Closes #<issue number here>.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*