Test times out
This test times out for me with the default 5 second timeout and a 10 second timeout. I bumped it to 60 seconds and it seemed to pass just fine. Without this I think a lot of peoples' builds will fail. This was on an M1 Pro with 16 GB of RAM so not the fastest machine, but not the slowest either.