chore(ec2): allow opting out of VPN Gateway lookup
When using `Vpc.fromLookup()`, a call may be made to determine whether a
VPN Gatway is attached. This call can be undesirable, for example when
the lookup role does not have permissions to make this call. This adds
a flag to allow opting out of VPN Gateways being queried to alleviate
this requirement.