chore: enforce `import/order` lint rule (#24254)
A recent change in a dependency we use to enforce the order of imports included a small update  to sort import from `./` before `../`.
We previously warned about violations of this rule, but did not enforce it.
This frequently causes files unrelated to a change being updated.

With this change I've updated all violations of this rule and the lint step is now free of warnings.
Most of these updates where automated, by passing `--fix` to the lint command.
For a few I had to manual resolve the problem.
In a small number of cases the import order could not be changed without causing a test failure.
For these I have just disabled the rule per line/block and did not investigate further.

This changes also updates the `import/order` rule to `error` to avoid any future issues.

Existing PRs will be fine, we enforce update from main anyway.
In some circumstances where new imports are added this will cause conflicts and require manual intervention.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*