fix(eks): changing the subnets or securityGroupIds order causes an error (#24163)
When the subnet list is passed to the EKS Cluster construct in a different order, an update is triggered to the EKS cluster. 
The update process fails as it falsely identifies a change for an unsupported update, although the list has the same items.

The solution is to change the analyzeUpdate function to return `updateVpc: false` if only the securityGroups/subnetsId order has been changed. 


Fixes: [#24162](https://github.com/aws/aws-cdk/issues/24162)

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license