feat(codebuild): adds file asset support to build-spec
Adds `fromAsset` to the `BuildSpec` class to allow referencing local
files as a project's buildspec. Uploads the file to s3 and references
the object arn in the `buildSpec` property of `Codebuild.Project`.

`isImmediate` is true for AssetBuildSpec because it's actual meaning is
not can be defined at synth time, but it exsists somewhere other  than
the project's source code, which in this case is true.

Requires referencing of the project so adds `scope` as an optional
parameter to `toBuildSpec` method.

fixes: #1138