feat(bootstrap): prevent accidental bootstrap overwrites
A problem that sometimes happens today is that an account has a
customized bootstrap stack setup, but a developer is not aware and they
`cdk bootstrap` the customizations away with the CDK default template.

This change introduces a new parameter, `BootstrapFlavor`, that
an organization can change to a different value. The CDK CLI will
refuse to overwrite existing bootstrap stacks of one flavor, with
a different one.

This string could have been in the template metadata, but making
it a parameter puts it front and center, making it more obvious
that template customizers should change it.

Does not change the bootstrap stack version itself, since nothing
about the resources inside changed.