fix(ecr-assets): fix repeated deploys of stacks with tar assets
The current implementation fails because it assumes the output of
`docker load` can only be

```
Loaded Image: {image}
```

However, if docker has to take any actions (such as untagging a previous
image) the output will look like

```
{error message}
{error message}
Loaded Image: {image}
```

To fix this, we can simply take the last line of the output via `tail`
before we attempt to extract the image. If the last line isn't of the
correct form, this will fail in an effectively equivalent way to how it
failed previously.

A previous attempt at fixing this (#18823) used a more complicated `sed`
command which used a cli flag that is not consistently available. This
approach won't work on all platforms (ex. macos) and is also much less
clear that it's correct.

This change also adds an integration test that tests the tarbell stack, though
it doesn't test the repeated deploy automatically. I have tested that by
running the integration test multiple times with an image modification in
between. We also vendor the tarball of the hello-world docker image, which
results in ~10kb of additional data in the repo. This is about as small as we
can get a vendored image with as straightforward a vendoring process.

fixes #18822