fix(secretsmanager): secret resource policy already exists in stack
The issue is prevalent with Database clusters, e.g. when a secret with a rotation is used and a grantee is granted access to the db cluster.

The root cause of this issue is the Cluster's use of a SecretAttachmentTarget to replace the original secret.
This is allowed, since SecretAttachmentTarget do implement the ISecret interface.
When policy statements are added, a new resource policy is created using the SecretAttachmentTarget ARN.
This only works because in CloudFormation SecretAttachmentTarget returns the ARN of the Secret.
However a secret can only have a single resource policy.
The above scenario leads to a policy being created for the Secret and one for the SecretAttachmentTarget.
Which is then (correctly) rejected by CloudFormation.

The fix is to forward any policy changes of the SecretAttachmentTarget to the attached secret.