feat(remodel): make tests more directory agnostic (#24359)
Adds a bunch of changes to various unit tests to make them work regardless of their location within the repository as long as they are relative to the files they reference. Before some of these depended on the current working directory being different than they will be in aws-cdk-lib. Some of these still need to get  update but will be easier to do by hand.

Additionally adds a bunch of missing build steps to `aws-cdk-lib` package.json.cdk-build.pre. These steps were all in individual modules before and built artifacts for various tests.

Also adds the `runBuild` function back into remodel. This builds everything, which unfortunately we need to do to be able to generate the alpha packages` individual pkg gen requires a bunch of the build tools and the packages that are generated themselves to actually be built first.