feat(remodel): make codegen handle cfn-include cfn2ts-cfn-include-map
Adds functionality to cfn2ts and aws-cdk-lib gen.ts to correctly
generate the map of cloudformation resource type names, to classes
within aws-cdk-lib. this was previously handled by a build script in
@aws-cdk/aws-cloudformation-include, but that depended on all of the
individual @aws-cdk/ packages being built and parsing the code using the
JSII assembly and jsii-reflect. This is not possible anymore because we
can't assure aws-cdk-lib is built before aws-cdk-lib is built.

However, this is pretty easy to do during code generation and in fact
results in less duplication of logic.