fix(secretsmanager): secret resource policy already exists in stack (under feature flag) (#24365)
The issue is prevalent with Database clusters, e.g. when a secret with a rotation is used and a grantee is granted access to the db cluster.

The root cause of this issue is the Cluster's use of a SecretAttachmentTarget to replace the original secret. This is allowed, since SecretAttachmentTarget do implement the ISecret interface. When policy statements are added, a new resource policy is created using the SecretAttachmentTarget ARN. This only works because in CloudFormation SecretAttachmentTarget returns the ARN of the Secret. However a secret can only have a single resource policy. The above scenario leads to a policy being created for the Secret and one for the SecretAttachmentTarget. Which is then (correctly) rejected by CloudFormation.

The fix is to forward any policy changes on the SecretAttachmentTarget to the attached secret.

Fixes #24383

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*