fix(cdk-assets): Error when building Docker Image Assets with Podman (#24003)
Encountering the same error as https://github.com/aws/aws-cdk/issues/16209#issue-978267269. When using `podman inspect` to check whether or not an image exists, the exit code for when an image does not exist is `125`, while Docker's is `1`. This change will treat either of these exit codes as meaning that the image with the given tag does not currently exist.

Closes #16209

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*