feat(apigateway): minCompressionSize on SpecRestApi (#24067)
This PR adds minCompressionSize to SpecRestApi, which allows compression on api's payload. Within SpecRestApi ``` minComppresssionSize: Size``` is the additional prop added to SpecRestApiProps to support minCompressionSize  which takes class Size, the value of how much the payload needs to be compressed. 

This PR  also adds minCompressionSize to RestApi which supports  Size class e.g ``` minComppresssionSize: Size``` and deprecates minimumCompressionSize which has number as a type.

For example,
```ts
const api = new apigw.RestApi(stack, 'RestApi', {
      minCompressionSize: Size.bytes(1024),
    });
```
compression for the payload would be 1024 bytes which is equivalent to 1 kibibyte.

Closes #22926.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*