chore: reduce amount of throttling to CodeArtifact (#24402)
Our new mechanism of testing candidate packages in our pipeline is by uploading them to CodeArtifact, instead of using hacky local tools that fake being NPM, NuGet, Maven, etc.

For the regression tests, we need access to versions `X` and `X+1` for packages, so we set all packages to `upstream: ALLOW` (individually, because there is no other way). This giant call volume is causing throttling which is alarming to the CodeArtifact team.

Only enable configuring upstreams for the regression test jobs, which is a lot less than for all tests.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*