fix(remodel): fix more issues with unit tests
Adds a bunch more replacements to unit tests to make them pass. Also
moves generating the region-info static data from `gen.ts` to a `pre`
step in `cdk-build` section of package.json. I'm not sure if this fixes
the issue but before when running gen for the first time it would
complain that the region-info gen function was not callable. This caused
some failures in region lookups of tests because region would be
undefined unexpectedly.