fix(lambda-nodejs): Required auto prefix of `handler` with `index.` breaks custom non-`index` handler settings used by layers (#24406)
Using `lambda-nodejs` makes it very easy to bundle functions with `esbuild`, but the code currently *always* prefixes the `handler` value with `index.`, which makes it impossible to use some lambda extensions together with this module as they require setting `handler` to specific values and the `index.` prefixing breaks the ability to set the handler to those values.

This PR avoids adding the `index.` prefix if the specified `handler` value contains a `.` already.

Closes #24403

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*