chore(cli-integ): ResourcePool will not reacquire dead processes
The ResourcePool acquires any available region from a collection of
regions, each protected by a mutex.

The underlying mutex supports stealing locks from processes that have
died, but the ResourcePool built on top never exercises that capability,
because it will only try to acquire locks from mutexes it "knows" are
free (i.e., for which it has seen an "unlock" event).

Try every mutex on every acquisition. This is less efficient, but it
has the advantage of recovering from processes that died.

There is no test for this as the test is very awkward to write (depends
on subprocesses being dead).

Also fix a bug in XpMutex itself where `acquire()` was supposed to
call `tryAcquire()`, but that was fat-fingered into a recursive call.